+++ src/plugins/intel_gpu/src/kernel_selector/kernels/gemm/gemm_kernel_base.h
@@ -22,8 +22,8 @@ struct gemm_params : public base_params {
     std::vector<int64_t> input0_order;
     std::vector<int64_t> input1_order;
     std::vector<int64_t> output_order;
-    int32_t input0_reshape_axes = 0;
-    int32_t input1_reshape_axes = 0;
+    int64_t input0_reshape_axes = 0;
+    int64_t input1_reshape_axes = 0;
     int64_t input0_broadcast_val = 0;
     int64_t input1_broadcast_val = 0;
     DataTensor beam_table;
